+++ src/fr/maxlego08/menu/loader/ZButtonLoader.java
@@ -11,6 +11,7 @@ import fr.maxlego08.menu.api.enums.PlaceholderAction;
 import fr.maxlego08.menu.api.event.events.ButtonLoadEvent;
 import fr.maxlego08.menu.api.loader.ButtonLoader;
 import fr.maxlego08.menu.api.loader.PermissibleLoader;
+import fr.maxlego08.menu.api.requirement.Action;
 import fr.maxlego08.menu.api.requirement.Requirement;
 import fr.maxlego08.menu.api.requirement.data.ActionPlayerData;
 import fr.maxlego08.menu.api.requirement.permissible.PlaceholderPermissible;
@@ -180,7 +181,7 @@ public class ZButtonLoader extends ZUtils implements Loader<Button> {
             }
         }
 
-        PermissibleLoader permissibleLoader = new PlaceholderPermissibleLoader();
+        PermissibleLoader permissibleLoader = new PlaceholderPermissibleLoader(this.plugin.getButtonManager());
         List<PlaceholderPermissible> placeholders = ((List<Map<String, Object>>) configuration.getList(path + "placeholders", new ArrayList<>())).stream().map(map -> {
             return (PlaceholderPermissible) permissibleLoader.load(path + "placeholders", new TypedMapAccessor(map), file);
         }).filter(permissible -> {
@@ -195,7 +196,7 @@ public class ZButtonLoader extends ZUtils implements Loader<Button> {
         PlaceholderAction placeholderAction = PlaceholderAction.from(configuration.getString(path + "action", null));
         String placeholderValue = configuration.getString(path + "value", null);
         if (placeholderAction != null && placeholderValue != null && placeholder != null) {
-            placeholders.add(new ZPlaceholderPermissible(placeholderAction, placeholder, placeholderValue));
+            placeholders.add(new ZPlaceholderPermissible(placeholderAction, placeholder, placeholderValue, new ArrayList<>(), new ArrayList<>()));
         }
 
         button.setPlaceholders(placeholders);
@@ -221,6 +222,9 @@ public class ZButtonLoader extends ZUtils implements Loader<Button> {
         loadViewRequirements(button, configuration, path, file);
         // Load clicks requirements
         loadClickRequirements(button, configuration, path, file);
+        // Load actions
+        List<Action> actions = buttonManager.loadActions((List<Map<String, Object>>) configuration.getList(path + "actions", new ArrayList<>()), path + "actions", file);
+        button.setActions(actions);
 
         InventoryManager inventoryManager = this.plugin.getInventoryManager();
         ButtonLoadEvent buttonLoadEvent = new ButtonLoadEvent(configuration, path, buttonManager, loader, button);