+++ src/fr/maxlego08/menu/requirement/permissible/ZRegexPermissible.java
@@ -1,21 +1,25 @@
 package fr.maxlego08.menu.requirement.permissible;
 
 import fr.maxlego08.menu.api.button.Button;
+import fr.maxlego08.menu.api.requirement.Action;
 import fr.maxlego08.menu.api.requirement.Permissible;
 import fr.maxlego08.menu.inventory.inventories.InventoryDefault;
+import fr.maxlego08.menu.requirement.ZPermissible;
 import fr.maxlego08.menu.zcore.logger.Logger;
 import fr.maxlego08.menu.zcore.utils.ZUtils;
 import jdk.jpackage.internal.Log;
 import org.bukkit.entity.Player;
 
+import java.util.List;
 import java.util.regex.Pattern;
 
-public class ZRegexPermissible extends ZUtils implements Permissible {
+public class ZRegexPermissible extends ZPermissible implements Permissible {
 
     private final Pattern pattern;
     private final String placeholder;
 
-    public ZRegexPermissible(String regex, String placeholder) {
+    public ZRegexPermissible(String regex, String placeholder, List<Action> denyActions, List<Action> successActions) {
+        super(denyActions, successActions);
         this.pattern = regex == null ? null : Pattern.compile(regex);
         this.placeholder = placeholder;
     }