+++ src/fr/maxlego08/menu/zcore/utils/PapiUtils.java
@@ -23,6 +23,8 @@ public class PapiUtils extends TranslationHelper {
      * @return string
      */
     public String papi(String placeHolder, Player player) {
+        // If the text does not contain a placeholder, then nothing is done
+        if (!placeHolder.contains("%")) return placeHolder;
         return this.use().setPlaceholders(player, placeHolder).replace("%player%", player.getName());
     }
 
@@ -34,6 +36,6 @@ public class PapiUtils extends TranslationHelper {
      * @return placeholders
      */
     public List<String> papi(List<String> placeHolders, Player player) {
-        return placeHolders.stream().map(e -> papi(e, player)).collect(Collectors.toList());
+        return placeHolders.stream().map(placeHolder -> papi(placeHolder, player)).collect(Collectors.toList());
     }
 }